% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_analysis}
\alias{add_analysis}
\title{Add analysis}
\usage{
add_analysis(
  analysis_ls = list(),
  mdls_ls,
  what_chr = c("wtp"),
  dce_design_ls = NULL,
  records_ls = NULL
)
}
\arguments{
\item{analysis_ls}{Analysis (a list), Default: list()}

\item{mdls_ls}{Models (a list)}

\item{what_chr}{What (a character vector), Default: c("wtp")}

\item{dce_design_ls}{Discrete choice experiment design (a list), Default: NULL}

\item{records_ls}{Records (a list), Default: NULL}
}
\value{
Analysis (a list)
}
\description{
add_analysis() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add analysis. The function returns Analysis (a list).
}
\keyword{internal}
