% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_choice_mdls}
\alias{add_choice_mdls}
\title{Add choice models}
\usage{
add_choice_mdls(
  mdls_ls = list(),
  dce_design_ls,
  mdl_params_ls,
  records_ls,
  exclude_chr = character(0),
  indl_predrs_chr = NA_character_,
  include_int = 1:4,
  max_concepts_1L_int = 2L,
  min_threshold_1L_int = 2L,
  nbr_of_clss_1L_int = 2L,
  purpose_chr = "attributes",
  significant_at_1L_dbl = 0.05,
  ...
)
}
\arguments{
\item{mdls_ls}{Models (a list), Default: list()}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{records_ls}{Records (a list)}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{indl_predrs_chr}{Individual predictors (a character vector), Default: 'NA'}

\item{include_int}{Include (an integer vector), Default: 1:4}

\item{max_concepts_1L_int}{Maximum concepts (an integer vector of length one), Default: 2}

\item{min_threshold_1L_int}{Minimum threshold (an integer vector of length one), Default: 2}

\item{nbr_of_clss_1L_int}{Number of classes (an integer vector of length one), Default: 2}

\item{purpose_chr}{Purpose (a character vector), Default: 'attributes'}

\item{significant_at_1L_dbl}{Significant at (a double vector of length one), Default: 0.05}

\item{...}{Additional arguments}
}
\value{
Models (a list)
}
\description{
add_choice_mdls() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add choice models. Function argument mdls_ls specifies the object to be updated. The function returns Models (a list).
}
\keyword{internal}
