% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cost_comparison}
\alias{add_cost_comparison}
\title{Add cost comparison}
\usage{
add_cost_comparison(
  analysis_ls,
  choices_ls,
  dce_design_ls,
  mdls_ls,
  mdl_params_ls,
  records_ls,
  altv_nms_chr = character(0),
  altv_to_modify_1L_int = 1,
  cost_range_dbl = 0:50,
  with_chr = "mnl_mlogit_mdl"
)
}
\arguments{
\item{analysis_ls}{Analysis (a list)}

\item{choices_ls}{Choices (a list)}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdls_ls}{Models (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{records_ls}{Records (a list)}

\item{altv_nms_chr}{Alternative names (a character vector), Default: character(0)}

\item{altv_to_modify_1L_int}{Alternative to modify (an integer vector of length one), Default: 1}

\item{cost_range_dbl}{Cost range (a double vector), Default: 0:50}

\item{with_chr}{With (a character vector), Default: 'mnl_mlogit_mdl'}
}
\value{
Analysis (a list)
}
\description{
add_cost_comparison() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add cost comparison. Function argument analysis_ls specifies the object to be updated. The function returns Analysis (a list).
}
