% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_country_red_flag}
\alias{add_country_red_flag}
\title{Add country red flag}
\usage{
add_country_red_flag(data_tb, lat_lng_var_nms_chr, valid_countries_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{lat_lng_var_nms_chr}{Lat longitude variable names (a character vector)}

\item{valid_countries_chr}{Valid countries (a character vector)}
}
\value{
Data (a tibble)
}
\description{
add_country_red_flag() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add country red flag. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
