% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_cut_pnts_cmprsn}
\alias{add_cut_pnts_cmprsn}
\title{Add cut points comparison}
\usage{
add_cut_pnts_cmprsn(
  prpn_cmprsns_ls = list(),
  cmprsn_nm_1L_chr,
  ds_tb,
  grouping_var_nms_chr,
  expected_dbl = numeric(0),
  is_pc_1L_lgl = T
)
}
\arguments{
\item{prpn_cmprsns_ls}{Proportion comparisons (a list), Default: list()}

\item{cmprsn_nm_1L_chr}{Comparison name (a character vector of length one)}

\item{ds_tb}{Dataset (a tibble)}

\item{grouping_var_nms_chr}{Grouping variable names (a character vector)}

\item{expected_dbl}{Expected (a double vector), Default: numeric(0)}

\item{is_pc_1L_lgl}{Is pc (a logical vector of length one), Default: T}
}
\value{
Proportion comparisons (a list)
}
\description{
add_cut_pnts_cmprsn() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add cut points comparison. The function returns Proportion comparisons (a list).
}
\keyword{internal}
