% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_design_spec}
\alias{add_design_spec}
\title{Add design specification}
\usage{
add_design_spec(
  dce_design_ls = list(),
  add_choice_cards_1L_lgl = F,
  add_cndt_design_mat = F,
  alternatives_chr = character(0),
  att_lvls_tb = tibble::tibble(),
  block_idxs_ls = list(),
  choice_var_pfx_1L_chr = "DCE_B",
  constraints_ls = list(),
  cost_att_idx_1L_int = integer(0),
  cost_pfx_1L_chr = "",
  cost_sfx_1L_chr = "",
  draws_1L_int = 10L,
  nbr_of_blocks_1L_int = integer(0),
  nbr_of_sets_1L_int = integer(0),
  opt_out_idx_1L_int = integer(0),
  parallel_1L_lgl = FALSE,
  pilot_ds_tb = NULL,
  priors_dbl = numeric(0),
  priors_idx_1L_int = integer(0),
  seed_1L_int = 1987,
  session_ls = list(),
  set_idx_1L_int = integer(0),
  transform_att_nms_1L_lgl = T
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list), Default: list()}

\item{add_choice_cards_1L_lgl}{Add choice cards (a logical vector of length one), Default: F}

\item{add_cndt_design_mat}{Add candidate design (a matrix), Default: F}

\item{alternatives_chr}{Alternatives (a character vector), Default: character(0)}

\item{att_lvls_tb}{Attribute levels (a tibble), Default: tibble::tibble()}

\item{block_idxs_ls}{Block indices (a list), Default: list()}

\item{choice_var_pfx_1L_chr}{Choice variable prefix (a character vector of length one), Default: 'DCE_B'}

\item{constraints_ls}{Constraints (a list), Default: list()}

\item{cost_att_idx_1L_int}{Cost attribute index (an integer vector of length one), Default: integer(0)}

\item{cost_pfx_1L_chr}{Cost prefix (a character vector of length one), Default: ''}

\item{cost_sfx_1L_chr}{Cost suffix (a character vector of length one), Default: ''}

\item{draws_1L_int}{Draws (an integer vector of length one), Default: 10}

\item{nbr_of_blocks_1L_int}{Number of blocks (an integer vector of length one), Default: integer(0)}

\item{nbr_of_sets_1L_int}{Number of sets (an integer vector of length one), Default: integer(0)}

\item{opt_out_idx_1L_int}{Opt out index (an integer vector of length one), Default: integer(0)}

\item{parallel_1L_lgl}{Parallel (a logical vector of length one), Default: FALSE}

\item{pilot_ds_tb}{Pilot dataset (a tibble), Default: NULL}

\item{priors_dbl}{Priors (a double vector), Default: numeric(0)}

\item{priors_idx_1L_int}{Priors index (an integer vector of length one), Default: integer(0)}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1987}

\item{session_ls}{Session (a list), Default: list()}

\item{set_idx_1L_int}{Set index (an integer vector of length one), Default: integer(0)}

\item{transform_att_nms_1L_lgl}{Transform attribute names (a logical vector of length one), Default: T}
}
\value{
Discrete choice experiment design (a list)
}
\description{
add_design_spec() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add design specification. Function argument dce_design_ls specifies the object to be updated. The function returns Discrete choice experiment design (a list).
}
\keyword{internal}
