% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_duplicate_text_red_flag}
\alias{add_duplicate_text_red_flag}
\title{Add duplicate text red flag}
\usage{
add_duplicate_text_red_flag(data_tb, unique_by_case_var_nms_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{unique_by_case_var_nms_chr}{Unique by case variable names (a character vector)}
}
\value{
Data (a tibble)
}
\description{
add_duplicate_text_red_flag() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add duplicate text red flag. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
