% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_email_red_flag}
\alias{add_email_red_flag}
\title{Add email red flag}
\usage{
add_email_red_flag(data_tb, email_var_nm_1L_chr, email_max_cnstv_digits_1L_int)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{email_var_nm_1L_chr}{Email variable name (a character vector of length one)}

\item{email_max_cnstv_digits_1L_int}{Email maximum consecutive digits (an integer vector of length one)}
}
\value{
Data (a tibble)
}
\description{
add_email_red_flag() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add email red flag. The function returns Data (a tibble).
}
\keyword{internal}
