% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_flags}
\alias{add_flags}
\title{Add flags}
\usage{
add_flags(
  records_ls,
  age_var_nm_1L_chr = character(0),
  attempt_dur_min_1L_dbl = numeric(0),
  attempt_dur_var_nm_1L_chr = character(0),
  date_stamp_format_1L_chr = "DMY",
  date_stamp_var_nm_1L_chr = character(0),
  dob_format_1L_chr = "DMY",
  dob_var_nm_1L_chr = character(0),
  email_max_cnstv_digits_1L_int = integer(0),
  email_var_nm_1L_chr = character(0),
  flags_max_1L_int = 0,
  flags_tot_var_nm_1L_chr = "red_flag_count_int",
  lat_lng_var_nms_chr = character(0),
  qltv_ax_green_flag_1L_chr = "Green",
  qltv_ax_red_flag_1L_chr = "Red",
  qltv_ax_tb = NULL,
  qltv_ax_var_nm_1L_chr = character(0),
  unique_by_case_var_nms_chr = character(0),
  valid_countries_chr = character(0)
)
}
\arguments{
\item{records_ls}{Records (a list)}

\item{age_var_nm_1L_chr}{Age variable name (a character vector of length one), Default: character(0)}

\item{attempt_dur_min_1L_dbl}{Attempt duration minimum (a double vector of length one), Default: numeric(0)}

\item{attempt_dur_var_nm_1L_chr}{Attempt duration variable name (a character vector of length one), Default: character(0)}

\item{date_stamp_format_1L_chr}{Date stamp format (a character vector of length one), Default: 'DMY'}

\item{date_stamp_var_nm_1L_chr}{Date stamp variable name (a character vector of length one), Default: character(0)}

\item{dob_format_1L_chr}{Date of birth format (a character vector of length one), Default: 'DMY'}

\item{dob_var_nm_1L_chr}{Date of birth variable name (a character vector of length one), Default: character(0)}

\item{email_max_cnstv_digits_1L_int}{Email maximum consecutive digits (an integer vector of length one), Default: integer(0)}

\item{email_var_nm_1L_chr}{Email variable name (a character vector of length one), Default: character(0)}

\item{flags_max_1L_int}{Flags maximum (an integer vector of length one), Default: 0}

\item{flags_tot_var_nm_1L_chr}{Flags total variable name (a character vector of length one), Default: 'red_flag_count_int'}

\item{lat_lng_var_nms_chr}{Lat longitude variable names (a character vector), Default: character(0)}

\item{qltv_ax_green_flag_1L_chr}{Qualitative assessment green flag (a character vector of length one), Default: 'Green'}

\item{qltv_ax_red_flag_1L_chr}{Qualitative assessment red flag (a character vector of length one), Default: 'Red'}

\item{qltv_ax_tb}{Qualitative assessment (a tibble), Default: NULL}

\item{qltv_ax_var_nm_1L_chr}{Qualitative assessment variable name (a character vector of length one), Default: character(0)}

\item{unique_by_case_var_nms_chr}{Unique by case variable names (a character vector), Default: character(0)}

\item{valid_countries_chr}{Valid countries (a character vector), Default: character(0)}
}
\value{
Records (a list)
}
\description{
add_flags() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add flags. The function returns Records (a list).
}
\keyword{internal}
