% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_nothing}
\alias{add_nothing}
\title{Add nothing}
\usage{
add_nothing()
}
\description{
add_nothing() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add nothing. The function is called for its side effects and does not return a value.
}
