% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_prpn_cmprsns}
\alias{add_prpn_cmprsns}
\title{Add proportion comparisons}
\usage{
add_prpn_cmprsns(
  comparison_tb,
  ds_tb,
  grouping_var_nms_chr,
  is_pc_1L_lgl = T,
  popl_var_nm_1L_chr = "Population (per-cent)",
  tmp_var_nm_1L_chr = "cmprsn_test_ls"
)
}
\arguments{
\item{comparison_tb}{Comparison (a tibble)}

\item{ds_tb}{Dataset (a tibble)}

\item{grouping_var_nms_chr}{Grouping variable names (a character vector)}

\item{is_pc_1L_lgl}{Is pc (a logical vector of length one), Default: T}

\item{popl_var_nm_1L_chr}{Population variable name (a character vector of length one), Default: 'Population (per-cent)'}

\item{tmp_var_nm_1L_chr}{Temporary variable name (a character vector of length one), Default: 'cmprsn_test_ls'}
}
\value{
Comparision (a tibble)
}
\description{
add_prpn_cmprsns() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add proportion comparisons. The function returns Comparision (a tibble).
}
\keyword{internal}
