% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_qltv_ax_red_flag}
\alias{add_qltv_ax_red_flag}
\title{Add qualitative assessment red flag}
\usage{
add_qltv_ax_red_flag(data_tb, qltv_ax_red_flag_1L_chr, qltv_ax_var_nm_1L_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{qltv_ax_red_flag_1L_chr}{Qualitative assessment red flag (a character vector of length one)}

\item{qltv_ax_var_nm_1L_chr}{Qualitative assessment variable name (a character vector of length one)}
}
\value{
Data (a tibble)
}
\description{
add_qltv_ax_red_flag() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add qualitative assessment red flag. The function returns Data (a tibble).
}
\keyword{internal}
