% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_red_flag_totals}
\alias{add_red_flag_totals}
\title{Add red flag totals}
\usage{
add_red_flag_totals(data_tb, flags_chr, flags_tot_var_nm_1L_chr)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{flags_chr}{Flags (a character vector)}

\item{flags_tot_var_nm_1L_chr}{Flags total variable name (a character vector of length one)}
}
\value{
Data (a tibble)
}
\description{
add_red_flag_totals() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add red flag totals. The function returns Data (a tibble).
}
\keyword{internal}
