% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_reported_age_red_flag}
\alias{add_reported_age_red_flag}
\title{Add reported age red flag}
\usage{
add_reported_age_red_flag(
  data_tb,
  date_stamp_var_nm_1L_chr,
  dob_var_nm_1L_chr,
  age_var_nm_1L_chr,
  date_stamp_format_1L_chr = "DMY",
  dob_format_1L_chr = "DMY"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{date_stamp_var_nm_1L_chr}{Date stamp variable name (a character vector of length one)}

\item{dob_var_nm_1L_chr}{Date of birth variable name (a character vector of length one)}

\item{age_var_nm_1L_chr}{Age variable name (a character vector of length one)}

\item{date_stamp_format_1L_chr}{Date stamp format (a character vector of length one), Default: 'DMY'}

\item{dob_format_1L_chr}{Date of birth format (a character vector of length one), Default: 'DMY'}
}
\value{
Data (a tibble)
}
\description{
add_reported_age_red_flag() is an Add function that updates an object by adding data to that object. Specifically, this function implements an algorithm to add reported age red flag. Function argument data_tb specifies the object to be updated. The function returns Data (a tibble).
}
\keyword{internal}
