% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_add.R
\name{add_sias_totals}
\alias{add_sias_totals}
\title{Add Social Interaction Anxiety Scale totals}
\usage{
add_sias_totals(
  data_tb,
  itm_prefix_1L_chr,
  drvd_var_prefix_1L_chr,
  sias_ctg_sfx_1L_chr = "SIAS_ctg",
  sias_tot_sfx_1L_chr = "SIAS_tot"
)
}
\arguments{
\item{data_tb}{Data (a tibble)}

\item{itm_prefix_1L_chr}{Item prefix (a character vector of length one)}

\item{drvd_var_prefix_1L_chr}{Derived variable prefix (a character vector of length one)}

\item{sias_ctg_sfx_1L_chr}{Social Interaction Anxiety Scale category suffix (a character vector of length one), Default: 'SIAS_ctg'}

\item{sias_tot_sfx_1L_chr}{Social Interaction Anxiety Scale total suffix (a character vector of length one), Default: 'SIAS_tot'}
}
\value{
Data (a tibble)
}
\description{
add_sias_totals() is an Add function that updates an object by adding new values to new or empty fields. Specifically, this function implements an algorithm to add social interaction anxiety scale totals. The function returns Data (a tibble).
}
\keyword{internal}
