% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_fit.R
\name{fit_choice_mdl}
\alias{fit_choice_mdl}
\title{Fit choice model}
\usage{
fit_choice_mdl(
  dce_design_ls,
  records_ls,
  mdl_params_ls,
  return_1L_chr = "mnl",
  correlation_1L_lgl = F,
  formula_env = new.env(parent = globalenv()),
  halton_xx = NA,
  indl_predrs_chr = NA_character_,
  mvar_ls = NULL,
  nbr_of_clss_1L_int = 2L,
  smnl_mdl = NULL,
  use_mlogit_pkg_1L_lgl = F,
  ...
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{records_ls}{Records (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'mnl'}

\item{correlation_1L_lgl}{Correlation (a logical vector of length one), Default: F}

\item{formula_env}{Formula (an environment), Default: new.env(parent = globalenv())}

\item{halton_xx}{Halton (an output object of multiple potential types), Default: NA}

\item{indl_predrs_chr}{Individual predictors (a character vector), Default: 'NA'}

\item{mvar_ls}{Variables that enter the mean of random parameters (a list), Default: NULL}

\item{nbr_of_clss_1L_int}{Number of classes (an integer vector of length one), Default: 2}

\item{smnl_mdl}{Scaled Multinomial Logit Model (a model), Default: NULL}

\item{use_mlogit_pkg_1L_lgl}{Use mlogit package package (a logical vector of length one), Default: F}

\item{...}{Additional arguments}
}
\value{
Model (a model)
}
\description{
fit_choice_mdl() is a Fit function that fits a model of a specified type to a dataset. Specifically, this function implements an algorithm to fit choice model. The function returns Model (a model).
}
\keyword{internal}
