% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_att_smrys}
\alias{get_att_smrys}
\title{Get attribute summarys}
\usage{
get_att_smrys(dce_design_ls, return_1L_chr = "levels")
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'levels'}
}
\value{
Return (an output object of multiple potential types)
}
\description{
get_att_smrys() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get attribute summarys. The function returns Return (an output object of multiple potential types).
}
\keyword{internal}
