% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_atts}
\alias{get_atts}
\title{Get attributes}
\usage{
get_atts(att_lvls_tb, return_1L_chr = "all")
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
Attributes (a character vector)
}
\description{
get_atts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get attributes. Function argument att_lvls_tb specifies the where to look for the required object. The function returns Attributes (a character vector).
}
\keyword{internal}
