% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_fctr_atts_dummy_var_nms}
\alias{get_fctr_atts_dummy_var_nms}
\title{Get factor attributes dummy variable names}
\usage{
get_fctr_atts_dummy_var_nms(att_lvls_tb, flatten_1L_lgl = F)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{flatten_1L_lgl}{Flatten (a logical vector of length one), Default: F}
}
\value{
Factor attributes dummy variable names (an output object of multiple potential types)
}
\description{
get_fctr_atts_dummy_var_nms() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get factor attributes dummy variable names. The function returns Factor attributes dummy variable names (an output object of multiple potential types).
}
\keyword{internal}
