% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_lvls}
\alias{get_lvls}
\title{Get levels}
\usage{
get_lvls(att_lvls_tb, return_1L_chr = "all")
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
Levels (a list)
}
\description{
get_lvls() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get levels. Function argument att_lvls_tb specifies the where to look for the required object. The function returns Levels (a list).
}
\keyword{internal}
