% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_opt_out_var_nm}
\alias{get_opt_out_var_nm}
\title{Get opt out variable name}
\usage{
get_opt_out_var_nm(case_choices_mat, choice_sets_ls)
}
\arguments{
\item{case_choices_mat}{Case choices (a matrix)}

\item{choice_sets_ls}{Choice sets (a list)}
}
\value{
Opt out name (a character vector of length one)
}
\description{
get_opt_out_var_nm() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get opt out variable name. The function returns Opt out name (a character vector of length one).
}
\keyword{internal}
