% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_signft_concepts}
\alias{get_signft_concepts}
\title{Get significant concepts}
\usage{
get_signft_concepts(
  att_predn_mdls_ls,
  mdl_params_ls,
  min_threshold_1L_int = 1L,
  exclude_chr = character(0)
)
}
\arguments{
\item{att_predn_mdls_ls}{Attribute prediction models (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{min_threshold_1L_int}{Minimum threshold (an integer vector of length one), Default: 1}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}
}
\value{
Significant concepts (a character vector)
}
\description{
get_signft_concepts() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get significant concepts. Function argument att_predn_mdls_ls specifies the where to look for the required object. The function returns Significant concepts (a character vector).
}
\keyword{internal}
