% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_knit.R
\name{knit_choice_descvs_rprt}
\alias{knit_choice_descvs_rprt}
\title{Knit choice descriptives report}
\usage{
knit_choice_descvs_rprt(choice_descvs_params_ls, path_to_mdl_rprt_tmpl_1L_chr)
}
\arguments{
\item{choice_descvs_params_ls}{Choice descriptives parameters (a list)}

\item{path_to_mdl_rprt_tmpl_1L_chr}{Path to model report template (a character vector of length one)}
}
\description{
knit_choice_descvs_rprt() is a Knit function that knits a rmarkdown file Specifically, this function implements an algorithm to knit choice descriptives report. The function is called for its side effects and does not return a value.
}
\keyword{internal}
