% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_knit.R
\name{knit_indl_mdls_rprt}
\alias{knit_indl_mdls_rprt}
\title{Knit individual models report}
\usage{
knit_indl_mdls_rprt(indl_mdl_knit_params_ls, path_to_mdl_rprt_tmpl_1L_chr)
}
\arguments{
\item{indl_mdl_knit_params_ls}{Individual model knit parameters (a list)}

\item{path_to_mdl_rprt_tmpl_1L_chr}{Path to model report template (a character vector of length one)}
}
\description{
knit_indl_mdls_rprt() is a Knit function that knits a rmarkdown file Specifically, this function implements an algorithm to knit individual models report. The function is called for its side effects and does not return a value.
}
\keyword{internal}
