% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_launch.R
\name{launch_survey_preview}
\alias{launch_survey_preview}
\title{Launch survey preview}
\usage{
launch_survey_preview(
  dce_design_ls,
  block_1L_int = integer(0),
  button_txt_1L_chr = character(0),
  end_txt_1L_chr = character(0),
  intro_text_1L_chr = character(0),
  set_idx_1L_int = integer(0),
  transform_att_nms_1L_lgl = T
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{block_1L_int}{Block (an integer vector of length one), Default: integer(0)}

\item{button_txt_1L_chr}{Button text (a character vector of length one), Default: character(0)}

\item{end_txt_1L_chr}{End text (a character vector of length one), Default: character(0)}

\item{intro_text_1L_chr}{Intro text (a character vector of length one), Default: character(0)}

\item{set_idx_1L_int}{Set index (an integer vector of length one), Default: integer(0)}

\item{transform_att_nms_1L_lgl}{Transform attribute names (a logical vector of length one), Default: T}
}
\value{
No return value, called for side effects.
}
\description{
launch_survey_preview() is a Launch function that launches an R Shiny app. Specifically, this function implements an algorithm to launch survey preview. The function is called for its side effects and does not return a value.
}
\keyword{internal}
