% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_areas_lup}
\alias{make_areas_lup}
\title{Make areas lookup table}
\usage{
make_areas_lup(
  X_Ready4useRepos = ready4use::Ready4useRepos(),
  area_types_chr = c("POA", "SOS"),
  consent_1L_chr = "",
  fl_nm_1L_chr = "geo_import",
  matches_chr = character(0),
  match_var_nm_1L_chr = "POA_CODE16",
  repo_fl_fmt_1L_chr = ".tab",
  save_fmt_1L_chr = ".csv",
  save_type_1L_chr = "original",
  write_to_1L_chr = character(0),
  year_1L_int = 2016
)
}
\arguments{
\item{X_Ready4useRepos}{PARAM_DESCRIPTION, Default: ready4use::Ready4useRepos()}

\item{area_types_chr}{Area types (a character vector), Default: c("POA", "SOS")}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'geo_import'}

\item{matches_chr}{Matches (a character vector), Default: character(0)}

\item{match_var_nm_1L_chr}{Match variable name (a character vector of length one), Default: 'POA_CODE16'}

\item{repo_fl_fmt_1L_chr}{Repository file format (a character vector of length one), Default: '.tab'}

\item{save_fmt_1L_chr}{Save format (a character vector of length one), Default: '.csv'}

\item{save_type_1L_chr}{Save type (a character vector of length one), Default: 'original'}

\item{write_to_1L_chr}{Write to (a character vector of length one), Default: character(0)}

\item{year_1L_int}{Year (an integer vector of length one), Default: 2016}
}
\value{
Areas (a lookup table)
}
\description{
make_areas_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make areas lookup table. The function returns Areas (a lookup table).
}
\keyword{internal}
