% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_att_predn_mdls_ls}
\alias{make_att_predn_mdls_ls}
\title{Make attribute prediction models list}
\usage{
make_att_predn_mdls_ls(
  dce_design_ls,
  mdl_params_ls,
  records_ls,
  return_1L_chr = "mixl",
  formula_env = new.env(parent = globalenv()),
  significant_at_1L_dbl = 0.05
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{records_ls}{Records (a list)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'mixl'}

\item{formula_env}{Formula (an environment), Default: new.env(parent = globalenv())}

\item{significant_at_1L_dbl}{Significant at (a double vector of length one), Default: 0.05}
}
\value{
Return (a list)
}
\description{
make_att_predn_mdls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make attribute prediction models list. The function returns Return (a list).
}
\keyword{internal}
