% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_candidate_choice_predrs_ls}
\alias{make_candidate_choice_predrs_ls}
\title{Make candidate choice predictors list}
\usage{
make_candidate_choice_predrs_ls(dce_design_ls, mdl_params_ls, records_ls)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{records_ls}{Records (a list)}
}
\value{
Candidate choice predictors (a list)
}
\description{
make_candidate_choice_predrs_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make candidate choice predictors list. The function returns Candidate choice predictors (a list).
}
\keyword{internal}
