% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_candidate_predrs_chr}
\alias{make_candidate_predrs_chr}
\title{Make candidate predictors character vector}
\usage{
make_candidate_predrs_chr(
  candidate_predrs_tb,
  concepts_chr = character(0),
  ds_tb = NULL,
  types_chr = character(0),
  as_selection_ls_1L_lgl = F
)
}
\arguments{
\item{candidate_predrs_tb}{Candidate predictors (a tibble)}

\item{concepts_chr}{Concepts (a character vector), Default: character(0)}

\item{ds_tb}{Dataset (a tibble), Default: NULL}

\item{types_chr}{Types (a character vector), Default: character(0)}

\item{as_selection_ls_1L_lgl}{As selection list (a logical vector of length one), Default: F}
}
\value{
Candidate predictors (a character vector)
}
\description{
make_candidate_predrs_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make candidate predictors character vector. The function returns Candidate predictors (a character vector).
}
\keyword{internal}
