% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cards_html_ls}
\alias{make_cards_html_ls}
\title{Make cards html list}
\usage{
make_cards_html_ls(block_choice_tbs_ls)
}
\arguments{
\item{block_choice_tbs_ls}{Block choice tibbles (a list)}
}
\description{
make_cards_html_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cards html list. The function is called for its side effects and does not return a value.
}
\keyword{internal}
