% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_case_choices_ds}
\alias{make_case_choices_ds}
\title{Make case choices dataset}
\usage{
make_case_choices_ds(
  case_choices_mat,
  choice_sets_ls,
  new_opt_out_var_nm_1L_chr = "opt_out"
)
}
\arguments{
\item{case_choices_mat}{Case choices (a matrix)}

\item{choice_sets_ls}{Choice sets (a list)}

\item{new_opt_out_var_nm_1L_chr}{New opt out variable name (a character vector of length one), Default: 'opt_out'}
}
\value{
Case choices (a tibble)
}
\description{
make_case_choices_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make case choices dataset. The function returns Case choices (a tibble).
}
\keyword{internal}
