% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_atts}
\alias{make_choice_atts}
\title{Make choice attributes}
\usage{
make_choice_atts(choice_sets_ls, opt_out_var_nm_1L_chr = "opt_out")
}
\arguments{
\item{choice_sets_ls}{Choice sets (a list)}

\item{opt_out_var_nm_1L_chr}{Opt out variable name (a character vector of length one), Default: 'opt_out'}
}
\value{
Choice attributes (a character vector)
}
\description{
make_choice_atts() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice attributes. The function returns Choice attributes (a character vector).
}
\keyword{internal}
