% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_card_html}
\alias{make_choice_card_html}
\title{Make choice card html}
\usage{
make_choice_card_html(choice_card_tb)
}
\arguments{
\item{choice_card_tb}{Choice card (a tibble)}
}
\value{
Card (a kable)
}
\description{
make_choice_card_html() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice card html. The function is called for its side effects and does not return a value.
}
\seealso{
\code{\link[knitr:kable]{knitr::kable()}}
}
\keyword{internal}
