% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_cards_tb_ls}
\alias{make_choice_cards_tb_ls}
\title{Make choice cards tibble list}
\usage{
make_choice_cards_tb_ls(blocks_int, choices_tb)
}
\arguments{
\item{blocks_int}{Blocks (an integer vector)}

\item{choices_tb}{Choices (a tibble)}
}
\value{
Block choice tibbles (a list)
}
\description{
make_choice_cards_tb_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice cards tibble list. The function returns Block choice tibbles (a list).
}
\keyword{internal}
