% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_int}
\alias{make_choice_int}
\title{Make choice integer vector}
\usage{
make_choice_int(ds_tb, choice_sets_ls, choice_vars_pfx_1L_chr = "DCE_B")
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{choice_sets_ls}{Choice sets (a list)}

\item{choice_vars_pfx_1L_chr}{Choice variables prefix (a character vector of length one), Default: 'DCE_B'}
}
\value{
Choice (an integer vector)
}
\description{
make_choice_int() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice integer vector. The function returns Choice (an integer vector).
}
\keyword{internal}
