% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_mdlng_ds}
\alias{make_choice_mdlng_ds}
\title{Make choice modelling dataset}
\usage{
make_choice_mdlng_ds(
  case_choices_mat,
  candidate_predrs_tb,
  choice_sets_ls,
  ds_tb,
  person_uid_var_nm_1L_chr,
  alternative_var_nm_1L_chr = "alternative",
  as_selection_ls_1L_lgl = F,
  card_id_var_nm_1L_chr = "card_id",
  choice_vars_pfx_1L_chr = "DCE_B",
  concepts_chr = character(0),
  person_card_uid_var_nm_1L_chr = "PERSONcard_id",
  return_1L_chr = "dfidx",
  types_chr = character(0)
)
}
\arguments{
\item{case_choices_mat}{Case choices (a matrix)}

\item{candidate_predrs_tb}{Candidate predictors (a tibble)}

\item{choice_sets_ls}{Choice sets (a list)}

\item{ds_tb}{Dataset (a tibble)}

\item{person_uid_var_nm_1L_chr}{Person unique identifier variable name (a character vector of length one)}

\item{alternative_var_nm_1L_chr}{Alternative variable name (a character vector of length one), Default: 'alternative'}

\item{as_selection_ls_1L_lgl}{As selection list (a logical vector of length one), Default: F}

\item{card_id_var_nm_1L_chr}{Card identity variable name (a character vector of length one), Default: 'card_id'}

\item{choice_vars_pfx_1L_chr}{Choice variables prefix (a character vector of length one), Default: 'DCE_B'}

\item{concepts_chr}{Concepts (a character vector), Default: character(0)}

\item{person_card_uid_var_nm_1L_chr}{Person card unique identifier variable name (a character vector of length one), Default: 'PERSONcard_id'}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'dfidx'}

\item{types_chr}{Types (a character vector), Default: character(0)}
}
\value{
Choice modelling (an output object of multiple potential types)
}
\description{
make_choice_mdlng_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice modelling dataset. The function returns Choice modelling (an output object of multiple potential types).
}
\keyword{internal}
