% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_mdlng_fml}
\alias{make_choice_mdlng_fml}
\title{Make choice modelling formula}
\usage{
make_choice_mdlng_fml(
  choice_sets_ls,
  candidate_predrs_chr = character(0),
  choice_atts_chr = character(0),
  choice_var_nm_1L_chr = "choice",
  cost_var_nm_1L_chr = "Cost",
  formula_env = new.env(parent = globalenv()),
  opt_out_var_nm_1L_chr = "opt_out",
  return_1L_chr = "mnl",
  use_mlogit_pkg_1L_lgl = F,
  wtp_mdl_1L_lgl = F
)
}
\arguments{
\item{choice_sets_ls}{Choice sets (a list)}

\item{candidate_predrs_chr}{Candidate predictors (a character vector), Default: character(0)}

\item{choice_atts_chr}{Choice attributes (a character vector), Default: character(0)}

\item{choice_var_nm_1L_chr}{Choice variable name (a character vector of length one), Default: 'choice'}

\item{cost_var_nm_1L_chr}{Cost variable name (a character vector of length one), Default: 'Cost'}

\item{formula_env}{Formula (an environment), Default: new.env(parent = globalenv())}

\item{opt_out_var_nm_1L_chr}{Opt out variable name (a character vector of length one), Default: 'opt_out'}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'mnl'}

\item{use_mlogit_pkg_1L_lgl}{Use mlogit package package (a logical vector of length one), Default: F}

\item{wtp_mdl_1L_lgl}{Willingness to pay model (a logical vector of length one), Default: F}
}
\value{
Formula (formula)
}
\description{
make_choice_mdlng_fml() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice modelling formula. The function is called for its side effects and does not return a value.
}
\keyword{internal}
