% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_responses_ds}
\alias{make_choice_responses_ds}
\title{Make choice responses dataset}
\usage{
make_choice_responses_ds(ds_tb, choice_vars_pfx_1L_chr = "DCE_B")
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{choice_vars_pfx_1L_chr}{Choice variables prefix (a character vector of length one), Default: 'DCE_B'}
}
\value{
Choice responses (a tibble)
}
\description{
make_choice_responses_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice responses dataset. The function returns Choice responses (a tibble).
}
\keyword{internal}
