% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choice_smrys}
\alias{make_choice_smrys}
\title{Make choice summarys}
\usage{
make_choice_smrys(
  dce_design_ls,
  mdl_params_ls,
  records_ls,
  text_size_1L_int = 14,
  wrapping_1L_int = 40
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{records_ls}{Records (a list)}

\item{text_size_1L_int}{Text size (an integer vector of length one), Default: 14}

\item{wrapping_1L_int}{Wrapping (an integer vector of length one), Default: 40}
}
\value{
Choice by attributes (a list)
}
\description{
make_choice_smrys() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choice summarys. The function returns Choice by attributes (a list).
}
\keyword{internal}
