% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_choices_ls}
\alias{make_choices_ls}
\title{Make choices list}
\usage{
make_choices_ls(
  dce_design_ls,
  block_idx_1L_int = 1L,
  by_altv_1L_lgl = T,
  card_idx_1L_int = 1L,
  new_choices_ls = NULL,
  set_idx_1L_int = integer(0)
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{block_idx_1L_int}{Block index (an integer vector of length one), Default: 1}

\item{by_altv_1L_lgl}{By alternative (a logical vector of length one), Default: T}

\item{card_idx_1L_int}{Card index (an integer vector of length one), Default: 1}

\item{new_choices_ls}{New choices (a list), Default: NULL}

\item{set_idx_1L_int}{Set index (an integer vector of length one), Default: integer(0)}
}
\value{
Choices (a list)
}
\description{
make_choices_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make choices list. The function returns Choices (a list).
}
\keyword{internal}
