% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cont_atts_indcs}
\alias{make_cont_atts_indcs}
\title{Make continuous attributes indices}
\usage{
make_cont_atts_indcs(att_lvls_tb)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}
}
\value{
Continuous attributes indices (an integer vector)
}
\description{
make_cont_atts_indcs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make continuous attributes indices. The function returns Continuous attributes indices (an integer vector).
}
\keyword{internal}
