% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cont_atts_rename_ls}
\alias{make_cont_atts_rename_ls}
\title{Make continuous attributes rename list}
\usage{
make_cont_atts_rename_ls(att_lvls_tb)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}
}
\value{
Continuous attributes rename (a list)
}
\description{
make_cont_atts_rename_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make continuous attributes rename list. The function returns Continuous attributes rename (a list).
}
\keyword{internal}
