% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cont_atts_tmp_var_nms}
\alias{make_cont_atts_tmp_var_nms}
\title{Make continuous attributes temporary variable names}
\usage{
make_cont_atts_tmp_var_nms(att_lvls_tb)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}
}
\value{
Continuous attributes temporary variable names (a character vector)
}
\description{
make_cont_atts_tmp_var_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make continuous attributes temporary variable names. The function returns Continuous attributes temporary variable names (a character vector).
}
\keyword{internal}
