% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_cut_pnts_cmprsn}
\alias{make_cut_pnts_cmprsn}
\title{Make cut points comparison}
\usage{
make_cut_pnts_cmprsn(
  ds_tb,
  grouping_var_nms_chr,
  expected_dbl = numeric(0),
  is_pc_1L_lgl = T
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{grouping_var_nms_chr}{Grouping variable names (a character vector)}

\item{expected_dbl}{Expected (a double vector), Default: numeric(0)}

\item{is_pc_1L_lgl}{Is pc (a logical vector of length one), Default: T}
}
\value{
Comparison (a tibble)
}
\description{
make_cut_pnts_cmprsn() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make cut points comparison. The function returns Comparison (a tibble).
}
\keyword{internal}
