% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_dsn_mat_nms}
\alias{make_dsn_mat_nms}
\title{Make design matrix names}
\usage{
make_dsn_mat_nms(
  dce_design_ls,
  grouping_1L_int = integer(0),
  idx_1L_int = 1L,
  opt_out_nm_1L_chr = "Opt out"
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{grouping_1L_int}{Grouping (an integer vector of length one), Default: integer(0)}

\item{idx_1L_int}{Index (an integer vector of length one), Default: 1}

\item{opt_out_nm_1L_chr}{Opt out name (a character vector of length one), Default: 'Opt out'}
}
\value{
Rename (a character vector)
}
\description{
make_dsn_mat_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make design matrix names. The function returns Rename (a character vector).
}
\keyword{internal}
