% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_efnt_dsn_mat}
\alias{make_efnt_dsn_mat}
\title{Make efficient design matrix}
\usage{
make_efnt_dsn_mat(
  dce_design_ls,
  parallel_1L_lgl = FALSE,
  pilot_analysis_ls = list(),
  priors_idx_1L_int = 1L,
  set_idx_1L_int = 1L,
  start_dsn_mat_ls = NULL
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{parallel_1L_lgl}{Parallel (a logical vector of length one), Default: FALSE}

\item{pilot_analysis_ls}{Pilot analysis (a list), Default: list()}

\item{priors_idx_1L_int}{Priors index (an integer vector of length one), Default: 1}

\item{set_idx_1L_int}{Set index (an integer vector of length one), Default: 1}

\item{start_dsn_mat_ls}{Start design (a list of matrices), Default: NULL}
}
\value{
Efficient design (a matrix)
}
\description{
make_efnt_dsn_mat() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make efficient design matrix. The function returns Efficient design (a matrix).
}
\keyword{internal}
