% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fctr_atts_dummy_var_nms}
\alias{make_fctr_atts_dummy_var_nms}
\title{Make factor attributes dummy variable names}
\usage{
make_fctr_atts_dummy_var_nms(att_lvls_tb, flatten_1L_lgl = T)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{flatten_1L_lgl}{Flatten (a logical vector of length one), Default: T}
}
\value{
New dummy attribute variable name (an output object of multiple potential types)
}
\description{
make_fctr_atts_dummy_var_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make factor attributes dummy variable names. The function returns New dummy attribute variable name (an output object of multiple potential types).
}
\keyword{internal}
