% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fctr_atts_indcs}
\alias{make_fctr_atts_indcs}
\title{Make factor attributes indices}
\usage{
make_fctr_atts_indcs(att_lvls_tb)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}
}
\value{
Factor attributes indices (an integer vector)
}
\description{
make_fctr_atts_indcs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make factor attributes indices. The function returns Factor attributes indices (an integer vector).
}
\keyword{internal}
