% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_fctr_atts_tmp_var_nms}
\alias{make_fctr_atts_tmp_var_nms}
\title{Make factor attributes temporary variable names}
\usage{
make_fctr_atts_tmp_var_nms(att_lvls_tb, flatten_1L_lgl = T, pfx_1L_chr = "Var")
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{flatten_1L_lgl}{Flatten (a logical vector of length one), Default: T}

\item{pfx_1L_chr}{Prefix (a character vector of length one), Default: 'Var'}
}
\value{
Factor attributes temporary variable names (an output object of multiple potential types)
}
\description{
make_fctr_atts_tmp_var_nms() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make factor attributes temporary variable names. The function returns Factor attributes temporary variable names (an output object of multiple potential types).
}
\keyword{internal}
