% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_flags_smry_ls}
\alias{make_flags_smry_ls}
\title{Make flags summary list}
\usage{
make_flags_smry_ls(
  ds_tb,
  qltv_ax_green_flag_1L_chr,
  qltv_ax_var_nm_1L_chr,
  flags_max_1L_int = 0,
  flags_tot_var_nm_1L_chr = "red_flag_count_int",
  preprocessing_log_ls = NULL
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{qltv_ax_green_flag_1L_chr}{Qualitative assessment green flag (a character vector of length one)}

\item{qltv_ax_var_nm_1L_chr}{Qualitative assessment variable name (a character vector of length one)}

\item{flags_max_1L_int}{Flags maximum (an integer vector of length one), Default: 0}

\item{flags_tot_var_nm_1L_chr}{Flags total variable name (a character vector of length one), Default: 'red_flag_count_int'}

\item{preprocessing_log_ls}{Preprocessing log (a list), Default: NULL}
}
\value{
Flags summary (a list)
}
\description{
make_flags_smry_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make flags summary list. The function returns Flags summary (a list).
}
\keyword{internal}
