% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_gmnl_mdl_smry}
\alias{make_gmnl_mdl_smry}
\title{Make generalised mutinomial logit model model summary}
\usage{
make_gmnl_mdl_smry(gmnl_mdl)
}
\arguments{
\item{gmnl_mdl}{Generalised mutinomial logit model (a model)}
}
\value{
Generalised mutinomial logit model summary (a list)
}
\description{
make_gmnl_mdl_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make generalised mutinomial logit model model summary. The function returns Generalised mutinomial logit model summary (a list).
}
\keyword{internal}
