% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_indl_mdl_smry_tb}
\alias{make_indl_mdl_smry_tb}
\title{Make individual model summary tibble}
\usage{
make_indl_mdl_smry_tb(mdls_ls)
}
\arguments{
\item{mdls_ls}{Models (a list)}
}
\value{
Individual model summary (a tibble)
}
\description{
make_indl_mdl_smry_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make individual model summary tibble. The function returns Individual model summary (a tibble).
}
\keyword{internal}
