% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_indl_mdls_knit_params}
\alias{make_indl_mdls_knit_params}
\title{Make individual models knit parameters}
\usage{
make_indl_mdls_knit_params(mdls_ls)
}
\arguments{
\item{mdls_ls}{Models (a list)}
}
\value{
Individual models knit parameters (a list)
}
\description{
make_indl_mdls_knit_params() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make individual models knit parameters. The function returns Individual models knit parameters (a list).
}
\keyword{internal}
