% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_params_ls}
\alias{make_mdl_params_ls}
\title{Make model parameters list}
\usage{
make_mdl_params_ls(
  candidate_predrs_tb,
  dce_design_ls,
  records_ls,
  as_selection_ls_1L_lgl = F,
  concepts_chr = character(0),
  draws_1L_int = 50L,
  opt_out_dstr_1L_chr = "n",
  types_chr = character(0)
)
}
\arguments{
\item{candidate_predrs_tb}{Candidate predictors (a tibble)}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{records_ls}{Records (a list)}

\item{as_selection_ls_1L_lgl}{As selection list (a logical vector of length one), Default: F}

\item{concepts_chr}{Concepts (a character vector), Default: character(0)}

\item{draws_1L_int}{Draws (an integer vector of length one), Default: 50}

\item{opt_out_dstr_1L_chr}{Opt out distribution (a character vector of length one), Default: 'n'}

\item{types_chr}{Types (a character vector), Default: character(0)}
}
\value{
Model parameters (a list)
}
\description{
make_mdl_params_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model parameters list. The function returns Model parameters (a list).
}
\keyword{internal}
