% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mdl_smry}
\alias{make_mdl_smry}
\title{Make model summary}
\usage{
make_mdl_smry(model_mdl, return_1L_chr = "coefficients")
}
\arguments{
\item{model_mdl}{Model (a model)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'coefficients'}
}
\value{
Return (an output object of multiple potential types)
}
\description{
make_mdl_smry() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make model summary. The function returns Return (an output object of multiple potential types).
}
\keyword{internal}
