% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_mxd_lgt_mdl_ls}
\alias{make_mxd_lgt_mdl_ls}
\title{Make mixed logit model list}
\usage{
make_mxd_lgt_mdl_ls(
  att_predn_mdls_ls,
  dce_design_ls,
  mdl_params_ls,
  records_ls,
  exclude_chr = character(0),
  formula_env = new.env(parent = globalenv()),
  max_concepts_1L_int = 3L,
  min_threshold_1L_int = 1L,
  return_1L_chr = "mixl"
)
}
\arguments{
\item{att_predn_mdls_ls}{Attribute prediction models (a list)}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{records_ls}{Records (a list)}

\item{exclude_chr}{Exclude (a character vector), Default: character(0)}

\item{formula_env}{Formula (an environment), Default: new.env(parent = globalenv())}

\item{max_concepts_1L_int}{Maximum concepts (an integer vector of length one), Default: 3}

\item{min_threshold_1L_int}{Minimum threshold (an integer vector of length one), Default: 1}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'mixl'}
}
\value{
Mixed logit (a list of models)
}
\description{
make_mxd_lgt_mdl_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make mixed logit model list. The function returns Mixed logit (a list of models).
}
\keyword{internal}
