% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_nbr_of_lvls}
\alias{make_nbr_of_lvls}
\title{Make number of levels}
\usage{
make_nbr_of_lvls(att_lvls_tb, return_1L_chr = "all")
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
Number of levels (an integer vector)
}
\description{
make_nbr_of_lvls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make number of levels. The function returns Number of levels (an integer vector).
}
\keyword{internal}
