% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_number_text}
\alias{make_number_text}
\title{Make number text}
\usage{
make_number_text(number_1L_int, capitalize_1L_lgl = F, threshold_1L_int = 10L)
}
\arguments{
\item{number_1L_int}{Number (an integer vector of length one)}

\item{capitalize_1L_lgl}{Capitalize (a logical vector of length one), Default: F}

\item{threshold_1L_int}{Threshold (an integer vector of length one), Default: 10}
}
\value{
Text (a character vector of length one)
}
\description{
make_number_text() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make number text. The function returns Text (a character vector of length one).
}
\keyword{internal}
