% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_participants_ds}
\alias{make_participants_ds}
\title{Make participants dataset}
\usage{
make_participants_ds(
  ds_tb,
  candidate_predrs_chr,
  choice_sets_ls,
  person_uid_var_nm_1L_chr,
  cases_are_responses_1L_lgl = F
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{candidate_predrs_chr}{Candidate predictors (a character vector)}

\item{choice_sets_ls}{Choice sets (a list)}

\item{person_uid_var_nm_1L_chr}{Person unique identifier variable name (a character vector of length one)}

\item{cases_are_responses_1L_lgl}{Cases are responses (a logical vector of length one), Default: F}
}
\value{
Participants (a tibble)
}
\description{
make_participants_ds() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make participants dataset. The function returns Participants (a tibble).
}
\keyword{internal}
