% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_pilot_analysis_ls}
\alias{make_pilot_analysis_ls}
\title{Make pilot analysis list}
\usage{
make_pilot_analysis_ls(
  pilot_ds_tb,
  dce_design_ls,
  constraints_ls = list(),
  choice_var_pfx_1L_chr = "DCE_B",
  draws_1L_int = 100L,
  seed_1L_int = 1987,
  set_idx_1L_int = 1L
)
}
\arguments{
\item{pilot_ds_tb}{Pilot dataset (a tibble)}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{constraints_ls}{Constraints (a list), Default: list()}

\item{choice_var_pfx_1L_chr}{Choice variable prefix (a character vector of length one), Default: 'DCE_B'}

\item{draws_1L_int}{Draws (an integer vector of length one), Default: 100}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1987}

\item{set_idx_1L_int}{Set index (an integer vector of length one), Default: 1}
}
\value{
Pilot analysis (a list)
}
\description{
make_pilot_analysis_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make pilot analysis list. The function returns Pilot analysis (a list).
}
\keyword{internal}
