% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_priors_ls}
\alias{make_priors_ls}
\title{Make priors list}
\usage{
make_priors_ls(
  dce_design_ls,
  priors_dbl,
  draws_1L_int = 10L,
  seed_1L_int = 1987
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{priors_dbl}{Priors (a double vector)}

\item{draws_1L_int}{Draws (an integer vector of length one), Default: 10}

\item{seed_1L_int}{Seed (an integer vector of length one), Default: 1987}
}
\value{
Priors (a list)
}
\description{
make_priors_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make priors list. The function returns Priors (a list).
}
\keyword{internal}
