% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_random_params_chr}
\alias{make_random_params_chr}
\title{Make random parameters character vector}
\usage{
make_random_params_chr(
  att_lvls_tb,
  opt_out_dstr_1L_chr = character(0),
  opt_out_var_nm_1L_chr = "opt_out",
  return_1L_chr = "all"
)
}
\arguments{
\item{att_lvls_tb}{Attribute levels (a tibble)}

\item{opt_out_dstr_1L_chr}{Opt out distribution (a character vector of length one), Default: character(0)}

\item{opt_out_var_nm_1L_chr}{Opt out variable name (a character vector of length one), Default: 'opt_out'}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'all'}
}
\value{
Random parameters (a character vector)
}
\description{
make_random_params_chr() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make random parameters character vector. The function returns Random parameters (a character vector).
}
\keyword{internal}
