% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_records_ls}
\alias{make_records_ls}
\title{Make records list}
\usage{
make_records_ls(
  ds_tb,
  alternative_var_nm_1L_chr = "alternative",
  card_id_var_nm_1L_chr = "card_id",
  case_choices_mat = matrix(),
  choice_var_nm_1L_chr = "choice",
  choice_vars_pfx_1L_chr = "DCE_B",
  cost_var_nm_1L_chr = "Cost",
  ds_dfidx = NULL,
  flags_ls = list(),
  opt_out_var_nm_1L_chr = "opt_out",
  person_card_uid_var_nm_1L_chr = "PERSONcard_id",
  person_uid_var_nm_1L_chr = "person_uid"
)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{alternative_var_nm_1L_chr}{Alternative variable name (a character vector of length one), Default: 'alternative'}

\item{card_id_var_nm_1L_chr}{Card identity variable name (a character vector of length one), Default: 'card_id'}

\item{case_choices_mat}{Case choices (a matrix), Default: matrix()}

\item{choice_var_nm_1L_chr}{Choice variable name (a character vector of length one), Default: 'choice'}

\item{choice_vars_pfx_1L_chr}{Choice variables prefix (a character vector of length one), Default: 'DCE_B'}

\item{cost_var_nm_1L_chr}{Cost variable name (a character vector of length one), Default: 'Cost'}

\item{ds_dfidx}{Dataset (an indexed data frame), Default: NULL}

\item{flags_ls}{Flags (a list), Default: list()}

\item{opt_out_var_nm_1L_chr}{Opt out variable name (a character vector of length one), Default: 'opt_out'}

\item{person_card_uid_var_nm_1L_chr}{Person card unique identifier variable name (a character vector of length one), Default: 'PERSONcard_id'}

\item{person_uid_var_nm_1L_chr}{Person unique identifier variable name (a character vector of length one), Default: 'person_uid'}
}
\value{
Records (a list)
}
\description{
make_records_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make records list. The function returns Records (a list).
}
\keyword{internal}
