% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_seifa_lup}
\alias{make_seifa_lup}
\title{Make Socio-Economic Indices for Areas lookup table}
\usage{
make_seifa_lup(
  url_1L_chr = character(0),
  fl_nm_1L_chr = character(0),
  col_nms_chr = c("postcode_dbl", "score_dbl", "percentile_dbl"),
  consent_1L_chr = "",
  keep_cols_int = c(1, 3, 7),
  sheet_1L_int = 3L,
  percentile_var_nm_1L_chr = "percentile_dbl",
  quartile_var_nm_1L_chr = "quartile_dbl",
  range_1L_chr = "A6:Q2635",
  write_to_1L_chr = character(0)
)
}
\arguments{
\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: character(0)}

\item{col_nms_chr}{Column names (a character vector), Default: c("postcode_dbl", "score_dbl", "percentile_dbl")}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{keep_cols_int}{Keep columns (an integer vector), Default: c(1, 3, 7)}

\item{sheet_1L_int}{Sheet (an integer vector of length one), Default: 3}

\item{percentile_var_nm_1L_chr}{Percentile variable name (a character vector of length one), Default: 'percentile_dbl'}

\item{quartile_var_nm_1L_chr}{Quartile variable name (a character vector of length one), Default: 'quartile_dbl'}

\item{range_1L_chr}{Range (a character vector of length one), Default: 'A6:Q2635'}

\item{write_to_1L_chr}{Write to (a character vector of length one), Default: character(0)}
}
\value{
Socio-Economic Indices for Areas (a lookup table)
}
\description{
make_seifa_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make socio-economic indices for areas lookup table. The function returns Socio-Economic Indices for Areas (a lookup table).
}
\keyword{internal}
