% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_signft_concepts_tbl}
\alias{make_signft_concepts_tbl}
\title{Make significant concepts table}
\usage{
make_signft_concepts_tbl(att_predn_mdls_ls, mdl_params_ls)
}
\arguments{
\item{att_predn_mdls_ls}{Attribute prediction models (a list)}

\item{mdl_params_ls}{Model parameters (a list)}
}
\value{
Significant concepts (a tibble)
}
\description{
make_signft_concepts_tbl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make significant concepts table. The function returns Significant concepts (a tibble).
}
\keyword{internal}
