% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_grouping_idxs}
\alias{make_smry_grouping_idxs}
\title{Make summary grouping indices}
\usage{
make_smry_grouping_idxs(mdl_smry_tb, records_ls)
}
\arguments{
\item{mdl_smry_tb}{Model summary (a tibble)}

\item{records_ls}{Records (a list)}
}
\value{
Summary grouping indices (an integer vector)
}
\description{
make_smry_grouping_idxs() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary grouping indices. The function returns Summary grouping indices (an integer vector).
}
\keyword{internal}
