% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_smry_tb}
\alias{make_smry_tb}
\title{Make summary tibble}
\usage{
make_smry_tb(ds_tb, var_metadata_tb, digits_1L_int = 2L)
}
\arguments{
\item{ds_tb}{Dataset (a tibble)}

\item{var_metadata_tb}{Variable metadata (a tibble)}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 2}
}
\value{
Summary (a tibble)
}
\description{
make_smry_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make summary tibble. The function returns Summary (a tibble).
}
