% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_sos_lup}
\alias{make_sos_lup}
\title{Make sos lookup table}
\usage{
make_sos_lup(
  area_var_nm_1L_chr = "SOS_CODE_2016",
  fl_nm_1L_chr = "sos_data.zip",
  path_1L_chr = character(0),
  url_1L_chr = character(0),
  write_to_1L_chr = character(0)
)
}
\arguments{
\item{area_var_nm_1L_chr}{Area variable name (a character vector of length one), Default: 'SOS_CODE_2016'}

\item{fl_nm_1L_chr}{File name (a character vector of length one), Default: 'sos_data.zip'}

\item{path_1L_chr}{Path (a character vector of length one), Default: character(0)}

\item{url_1L_chr}{Url (a character vector of length one), Default: character(0)}

\item{write_to_1L_chr}{Write to (a character vector of length one), Default: character(0)}
}
\value{
Sos (a lookup table)
}
\description{
make_sos_lup() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make sos lookup table. The function returns Sos (a lookup table).
}
\keyword{internal}
