% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tfd_flags_smry_tb}
\alias{make_tfd_flags_smry_tb}
\title{Make transformed flags summary tibble}
\usage{
make_tfd_flags_smry_tb(
  preprocessing_log_ls,
  records_ls,
  country_1L_chr,
  what_1L_chr = "all"
)
}
\arguments{
\item{preprocessing_log_ls}{Preprocessing log (a list)}

\item{records_ls}{Records (a list)}

\item{country_1L_chr}{Country (a character vector of length one)}

\item{what_1L_chr}{What (a character vector of length one), Default: 'all'}
}
\value{
Transformed flags summary (a tibble)
}
\description{
make_tfd_flags_smry_tb() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transformed flags summary tibble. The function returns Transformed flags summary (a tibble).
}
\keyword{internal}
