% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_tfd_lvls_ls}
\alias{make_tfd_lvls_ls}
\title{Make transformed levels list}
\usage{
make_tfd_lvls_ls(dce_design_ls)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}
}
\value{
Levels (a list)
}
\description{
make_tfd_lvls_ls() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make transformed levels list. The function returns Levels (a list).
}
\keyword{internal}
