% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_make.R
\name{make_wtp_mat}
\alias{make_wtp_mat}
\title{Make willingness to pay matrix}
\usage{
make_wtp_mat(
  mnl_mdl,
  cost_var_nm_1L_chr = "Cost",
  digits_1L_int = 3L,
  return_1L_chr = "mat"
)
}
\arguments{
\item{mnl_mdl}{Multinomial logit (a model)}

\item{cost_var_nm_1L_chr}{Cost variable name (a character vector of length one), Default: 'Cost'}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 3}

\item{return_1L_chr}{Return (a character vector of length one), Default: 'mat'}
}
\value{
Willingness to pay (an output object of multiple potential types)
}
\description{
make_wtp_mat() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make willingness to pay matrix. The function returns Willingness to pay (an output object of multiple potential types).
}
\keyword{internal}
