% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_predict.R
\name{predict_mkt_share}
\alias{predict_mkt_share}
\title{Predict market share}
\usage{
predict_mkt_share(
  dce_design_ls,
  mdls_ls,
  mdl_params_ls,
  new_choices_ls,
  records_ls,
  altv_nms_chr = character(0),
  with_1L_chr = "mnl_mlogit_mdl"
)
}
\arguments{
\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdls_ls}{Models (a list)}

\item{mdl_params_ls}{Model parameters (a list)}

\item{new_choices_ls}{New choices (a list)}

\item{records_ls}{Records (a list)}

\item{altv_nms_chr}{Alternative names (a character vector), Default: character(0)}

\item{with_1L_chr}{With (a character vector of length one), Default: 'mnl_mlogit_mdl'}
}
\value{
Predicted share (a double vector)
}
\description{
predict_mkt_share() is a Predict function that makes predictions from data using a specified statistical model. Specifically, this function implements an algorithm to predict market share. The function returns Predicted share (a double vector).
}
\keyword{internal}
