% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_cmprsn_kbl}
\alias{print_cmprsn_kbl}
\title{Print comparison kable}
\usage{
print_cmprsn_kbl(cmprsn_tb, caption_1L_chr, footnotes_chr = character(0))
}
\arguments{
\item{cmprsn_tb}{Comparison (a tibble)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{footnotes_chr}{Footnotes (a character vector), Default: character(0)}
}
\value{
Comparison (a kable)
}
\description{
print_cmprsn_kbl() is a Print function that prints output to console Specifically, this function implements an algorithm to print comparison kable. The function is called for its side effects and does not return a value.
}
\keyword{internal}
