% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_mdl_smry_kbl}
\alias{print_mdl_smry_kbl}
\title{Print model summary kable}
\usage{
print_mdl_smry_kbl(
  tfd_mdl_smry_tb,
  records_ls,
  caption_1L_chr,
  footnotes_chr = character(0)
)
}
\arguments{
\item{tfd_mdl_smry_tb}{Transformed model summary (a tibble)}

\item{records_ls}{Records (a list)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{footnotes_chr}{Footnotes (a character vector), Default: character(0)}
}
\value{
Model summary (a kable)
}
\description{
print_mdl_smry_kbl() is a Print function that prints output to console Specifically, this function implements an algorithm to print model summary kable. The function is called for its side effects and does not return a value.
}
\keyword{internal}
