% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_print.R
\name{print_tfd_mdl_smry}
\alias{print_tfd_mdl_smry}
\title{Print transformed model summary}
\usage{
print_tfd_mdl_smry(
  model_mdl,
  caption_1L_chr,
  dce_design_ls,
  mdl_params_ls = NULL,
  records_ls,
  confidence_intvl_1L_chr = character(0),
  confidence_intvl_int = integer(0),
  digits_1L_int = 3L,
  footnotes_chr = character(0),
  print_call_1L_lgl = T,
  significance_1L_chr = character(0),
  opt_out_nm_1L_chr = "Opt out"
)
}
\arguments{
\item{model_mdl}{Model (a model)}

\item{caption_1L_chr}{Caption (a character vector of length one)}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{mdl_params_ls}{Model parameters (a list), Default: NULL}

\item{records_ls}{Records (a list)}

\item{confidence_intvl_1L_chr}{Confidence interval (a character vector of length one), Default: character(0)}

\item{confidence_intvl_int}{Confidence interval (an integer vector), Default: integer(0)}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: 3}

\item{footnotes_chr}{Footnotes (a character vector), Default: character(0)}

\item{print_call_1L_lgl}{Print call (a logical vector of length one), Default: T}

\item{significance_1L_chr}{Significance (a character vector of length one), Default: character(0)}

\item{opt_out_nm_1L_chr}{Opt out name (a character vector of length one), Default: 'Opt out'}
}
\value{
Transformed model summary (a kable)
}
\description{
print_tfd_mdl_smry() is a Print function that prints output to console. Specifically, this function implements an algorithm to print transformed model summary. The function is called for its side effects and does not return a value.
}
\keyword{internal}
