% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_remove.R
\name{remove_no_choice_responses}
\alias{remove_no_choice_responses}
\title{Remove no choice responses}
\usage{
remove_no_choice_responses(responses_tb, choice_card_pfx_1L_chr = "DCE_B")
}
\arguments{
\item{responses_tb}{Responses (a tibble)}

\item{choice_card_pfx_1L_chr}{Choice card prefix (a character vector of length one), Default: 'DCE_B'}
}
\value{
Responses (a tibble)
}
\description{
remove_no_choice_responses() is a Remove function that edits an object, removing a specified element or elements. Specifically, this function implements an algorithm to remove no choice responses. The function returns Responses (a tibble).
}
\keyword{internal}
