% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reorder.R
\name{reorder_design_mat}
\alias{reorder_design_mat}
\title{Reorder design matrix}
\usage{
reorder_design_mat(design_mat, block_indcs_ls, choice_sets_ls)
}
\arguments{
\item{design_mat}{Design (a matrix)}

\item{block_indcs_ls}{Block indices (a list)}

\item{choice_sets_ls}{Choice sets (a list)}
}
\value{
Reordered (a matrix)
}
\description{
reorder_design_mat() is a Reorder function that reorders an object to conform to a pre-specified schema. Specifically, this function implements an algorithm to reorder design matrix. The function returns Reordered (a matrix).
}
\keyword{internal}
