% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_descvs_tb}
\alias{transform_descvs_tb}
\title{Transform descriptives tibble}
\usage{
transform_descvs_tb(descvs_tb)
}
\arguments{
\item{descvs_tb}{Descriptives (a tibble)}
}
\value{
Transformed descriptives (a tibble)
}
\description{
transform_descvs_tb() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform descriptives tibble. The function returns Transformed descriptives (a tibble).
}
\keyword{internal}
