% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_mdl_smry}
\alias{transform_mdl_smry}
\title{Transform model summary}
\usage{
transform_mdl_smry(
  mdl_smry_tb,
  dce_design_ls,
  records_ls,
  mdl_params_ls = NULL,
  confidence_intvl_1L_chr = character(0),
  confidence_intvl_int = integer(0),
  digits_1L_int = integer(0),
  opt_out_nm_1L_chr = "Opt out",
  stat_1L_chr = "z"
)
}
\arguments{
\item{mdl_smry_tb}{Model summary (a tibble)}

\item{dce_design_ls}{Discrete choice experiment design (a list)}

\item{records_ls}{Records (a list)}

\item{mdl_params_ls}{Model parameters (a list), Default: NULL}

\item{confidence_intvl_1L_chr}{Confidence interval (a character vector of length one), Default: character(0)}

\item{confidence_intvl_int}{Confidence interval (an integer vector), Default: integer(0)}

\item{digits_1L_int}{Digits (an integer vector of length one), Default: integer(0)}

\item{opt_out_nm_1L_chr}{Opt out name (a character vector of length one), Default: 'Opt out'}

\item{stat_1L_chr}{Statistic (a character vector of length one), Default: 'z'}
}
\value{
Model summary (a tibble)
}
\description{
transform_mdl_smry() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform model summary. Function argument mdl_smry_tb specifies the object to be updated. Argument dce_design_ls provides the object to be updated. The function returns Model summary (a tibble).
}
\keyword{internal}
