% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_repln_ds_for_analysis}
\alias{transform_repln_ds_for_analysis}
\title{Transform replication dataset for analysis}
\usage{
transform_repln_ds_for_analysis(
  repln_ds_tb,
  areas_lup = NULL,
  consent_1L_chr = "",
  person_uid_var_nm_1L_chr = "person_uid",
  seifa_lup = NULL,
  set_idx_1L_int = 2L,
  write_to_1L_chr = character(0)
)
}
\arguments{
\item{repln_ds_tb}{Replication dataset (a tibble)}

\item{areas_lup}{Areas (a lookup table), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}

\item{person_uid_var_nm_1L_chr}{Person unique identifier variable name (a character vector of length one), Default: 'person_uid'}

\item{seifa_lup}{Socio-Economic Indices for Areas (a lookup table), Default: NULL}

\item{set_idx_1L_int}{Set index (an integer vector of length one), Default: 2}

\item{write_to_1L_chr}{Write to (a character vector of length one), Default: character(0)}
}
\value{
Transformed replication dataset (a tibble)
}
\description{
transform_repln_ds_for_analysis() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform replication dataset for analysis. Function argument repln_ds_tb specifies the object to be updated. Argument areas_lup provides the object to be updated. The function returns Transformed replication dataset (a tibble).
}
