% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_transform.R
\name{transform_to_excel_date_fmt}
\alias{transform_to_excel_date_fmt}
\title{Transform to excel date format}
\usage{
transform_to_excel_date_fmt(
  date_chr,
  format_1L_chr = "ymd HMS",
  index_date_1L_chr = "1899-12-30"
)
}
\arguments{
\item{date_chr}{Date (a character vector)}

\item{format_1L_chr}{Format (a character vector of length one), Default: 'ymd HMS'}

\item{index_date_1L_chr}{Index date (a character vector of length one), Default: '1899-12-30'}
}
\description{
transform_to_excel_date_fmt() is a Transform function that edits an object in such a way that core object attributes - e.g. shape, dimensions, elements, type - are altered. Specifically, this function implements an algorithm to transform to excel date format. Function argument date_chr specifies the object to be updated. Argument format_1L_chr provides the object to be updated. The function is called for its side effects and does not return a value.
}
\keyword{internal}
