% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_choice_mdlng_ws}
\alias{write_choice_mdlng_ws}
\title{Write choice modelling workspace}
\usage{
write_choice_mdlng_ws(paths_ls, consent_1L_chr = "")
}
\arguments{
\item{paths_ls}{Paths (a list)}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}
}
\value{
Paths (a list)
}
\description{
write_choice_mdlng_ws() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write choice modelling workspace. The function returns Paths (a list).
}
\keyword{internal}
