% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_write.R
\name{write_preprocessing_outp}
\alias{write_preprocessing_outp}
\title{Write preprocessing output}
\usage{
write_preprocessing_outp(
  paths_ls,
  mdl_params_ls = NULL,
  preprocessing_log_ls = NULL,
  records_ls = NULL,
  consent_1L_chr = ""
)
}
\arguments{
\item{paths_ls}{Paths (a list)}

\item{mdl_params_ls}{Model parameters (a list), Default: NULL}

\item{preprocessing_log_ls}{Preprocessing log (a list), Default: NULL}

\item{records_ls}{Records (a list), Default: NULL}

\item{consent_1L_chr}{Consent (a character vector of length one), Default: ''}
}
\value{
Paths (a list)
}
\description{
write_preprocessing_outp() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write preprocessing output. The function returns Paths (a list).
}
\keyword{internal}
