% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_nbr_of_choices}
\alias{get_nbr_of_choices}
\title{Get number of choices}
\usage{
get_nbr_of_choices(choice_sets_ls)
}
\arguments{
\item{choice_sets_ls}{Choice sets (a list)}
}
\value{
Number of choices (an integer vector of length one)
}
\description{
get_nbr_of_choices() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get number of choices. Function argument choice_sets_ls specifies the where to look for the required object. The function returns Number of choices (an integer vector of length one).
}
\keyword{internal}
