% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychoice-package.R, R/pkg_mychoice.R
\docType{package}
\name{mychoice-package}
\alias{mychoice-package}
\title{mychoice: Model Youth Choice Behaviours}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for modelling choice behaviours in youth mental health economic models. Supports standardised workflows for the design and analysis of Discrete Choice Experiments. This development version of the mychoice package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for modelling choice behaviours in youth mental health economic models. Supports standardised workflows for the design and analysis of Discrete Choice Experiments. This development version of the mychoice package has been made available as part of the process of testing and documenting the package. If you have any questions, please contact the authors (matthew.hamilton1@monash.edu).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/mychoice/}
  \item \url{https://github.com/ready4-dev/mychoice}
  \item \url{https://www.ready4-dev.com/}
}


Useful links:
\itemize{
  \item \url{https://ready4-dev.github.io/mychoice/}
  \item \url{https://github.com/ready4-dev/mychoice}
  \item \url{https://www.ready4-dev.com/}
}

}
\author{
\strong{Maintainer}: Matthew Hamilton \email{matthew.hamilton1@monash.edu} (\href{https://orcid.org/0000-0001-7407-9194}{ORCID}) [funder, copyright holder]

Other contributors:
\itemize{
  \item Orygen [copyright holder, funder]
}

}
\keyword{internal}
